--- src/domain/i18n/index.js
@@ -1,3 +1,0 @@
-import * as i18n from './useCases';
-
-export default i18n;
