--- src/domain/i18n/useCases/index.js
@@ -1,11 +1,0 @@
-import detectLocale from './detectLocale';
-import ensureIsSupported from './ensureIsSupported';
-import getText from './getText';
-import getSupportedLanguages from './getSupportedLanguages';
-
-export {
-  detectLocale,
-  ensureIsSupported,
-  getText,
-  getSupportedLanguages
-};
