--- src/domain/i18n/valueObjects/currentLocale.js
@@ -1,5 +1,0 @@
-import { defaultLocale } from '../useCases/detectLocale';
-
-export default {
-  locale: defaultLocale
-};
