+++ src/domain/index.js
@@ -1,10 +1,8 @@
 import compose from './compose';
 import certificates from './certificates';
 import events from './events';
-import i18n from './i18n';
 
 export default compose({
   certificates,
-  events,
-  i18n
+  events
 });
