+++ src/i18n/getSupportedLanguages.js
@@ -1,4 +1,10 @@
-import i18n from '../../../i18n';
+import en from './lang/en.json';
+import fr from './lang/fr.json';
+
+const i18n = {
+  en,
+  fr
+};
 
 export default function getSupportedLanguages () {
   return Object.keys(i18n);
