+++ src/i18n/index.js
@@ -1,7 +1,11 @@
-import en from './lang/en.json';
-import fr from './lang/fr.json';
+import detectLocale from './detectLocale';
+import ensureIsSupported from './ensureIsSupported';
+import getText from './getText';
+import getSupportedLanguages from './getSupportedLanguages';
 
-export default {
-  en,
-  fr
+export {
+  detectLocale,
+  ensureIsSupported,
+  getText,
+  getSupportedLanguages
 };
