+++ src/i18n/valueObjects/currentLocale.js
@@ -1,0 +1,5 @@
+import { defaultLocale } from '../detectLocale';
+
+export default {
+  locale: defaultLocale
+};
