+++ test/application/i18n/ensureIsSupported.spec.js
@@ -1,12 +1,13 @@
 import sinon from 'sinon';
-import domain from '../../../../../src/domain';
+import * as i18n from '../../../src/i18n/index';
+import ensureIsSupported from '../../../src/i18n/ensureIsSupported';
 
 describe('domain i18n ensureIsSupported use case test suite', function () {
   let supportedStub;
   let instance;
 
   beforeEach(function () {
-    supportedStub = sinon.stub(domain.i18n, 'getSupportedLanguages').returns(['en', 'es', 'fr']);
+    supportedStub = sinon.stub(i18n, 'getSupportedLanguages').returns(['en', 'es', 'fr']);
   });
 
   afterEach(function () {
@@ -16,20 +17,20 @@ describe('domain i18n ensureIsSupported use case test suite', function () {
 
   describe('given the set locale is supported', function () {
     it('should return the set locale', function () {
-      instance = domain.i18n.ensureIsSupported('es');
+      instance = ensureIsSupported('es');
       expect(instance).toBe('es');
     });
 
     describe('and of different case', () => {
       it('should return the set locale', () => {
-        instance = domain.i18n.ensureIsSupported('es-es');
+        instance = ensureIsSupported('es-es');
         expect(instance).toBe('es');
       });
     });
 
     describe('given the set locale is RFC 3066 language format', () => {
       it('should return the language tag that first matches', () => {
-        instance = domain.i18n.ensureIsSupported('fr');
+        instance = ensureIsSupported('fr');
         expect(instance).toBe('fr');
       });
     });
@@ -37,7 +38,7 @@ describe('domain i18n ensureIsSupported use case test suite', function () {
 
   describe('given the set locale is not supported', function () {
     it('should return the default locale', function () {
-      instance = domain.i18n.ensureIsSupported('az-az');
+      instance = ensureIsSupported('az-az');
       expect(instance).toBe('en');
     });
   });
