+++ test/application/i18n/getText.spec.js
@@ -1,17 +1,17 @@
-import currentLocale from '../../../../../src/domain/i18n/valueObjects/currentLocale';
-import domain from '../../../../../src/domain';
+import currentLocale from '../../../src/i18n/valueObjects/currentLocale';
+import getText from '../../../src/i18n/getText';
 
 describe('domain i18n getText use case test suite', function () {
   describe('given it is invoked without the group', function () {
     it('should return an empty string', function () {
-      const res = domain.i18n.getText();
+      const res = getText();
       expect(res).toBe('');
     });
   });
 
   describe('given it is invoked without the item', function () {
     it('should return an error', function () {
-      const res = domain.i18n.getText('group');
+      const res = getText('group');
       expect(res).toBe('');
     });
   });
@@ -20,7 +20,7 @@ describe('domain i18n getText use case test suite', function () {
     // TODO: is test relevant if we ensure the locale exists?
     xit('should return an error', function () {
       currentLocale.locale = 'fr-FR';
-      const res = domain.i18n.getText('group', 'item');
+      const res = getText('group', 'item');
       expect(res).toBe('[missing locale data]');
       currentLocale.locale = 'en-US';
     });
@@ -28,14 +28,14 @@ describe('domain i18n getText use case test suite', function () {
 
   describe('given the group data does not exist in the i18n data', function () {
     it('should return an error', function () {
-      const res = domain.i18n.getText('invalid-group', 'item');
+      const res = getText('invalid-group', 'item');
       expect(res).toBe('[missing locale group data]');
     });
   });
 
   describe('given the item does not exist in the i18n data', function () {
     it('should return an error', function () {
-      const res = domain.i18n.getText('errors', 'invalid-item');
+      const res = getText('errors', 'invalid-item');
       expect(res).toBe('[missing locale item data]');
     });
   });
